WIP: add sync nodes and collector nodes

+++ raspi/gpio-tracer/include/la_sigrok.h
@@ -13,7 +13,7 @@ struct channel_mode {
 };
 
 /* --- proto --- */
-int la_sigrok_init_instance(guint64 samplerate, const gchar* logpath, GVariant* channel_modes);
+int la_sigrok_init_instance(guint32 samplerate, const gchar* logpath, GVariant* channel_modes);
 int la_sigrok_run_instance();
 int la_sigrok_stop_instance();
